+++ src/main/java/com/example/plan/customer/service/CustomerService.java
@@ -11,16 +11,16 @@ import java.util.Map;
 public interface CustomerService {
     List<Map<String, Object>> findAllCustomers();
 
-    Map<String, Object>  findById(int id);
+    Map<String, Object>  findById(Long id);
 
     Map<String, Object> findCustomerByName(String firstName, String lastName, LocalDate birthday);
 
     ResponseEntity<ResponseMessage> addCustomer(Map<String, Object> requestMap);
 
-    ResponseEntity<ResponseMessage> updateCustomer(Map<String, Object> requestMap, int id);
+    ResponseEntity<ResponseMessage> updateCustomer(Map<String, Object> requestMap, Long id);
 
-    ResponseEntity<ResponseMessage> deleteCustomerAndPlanById(int  customerId, int PlanId);
+    ResponseEntity<ResponseMessage> deleteCustomerAndPlanById(Long  customerId, int PlanId);
 
-    ResponseEntity<ResponseMessage> deleteCustomer(int id);
+    ResponseEntity<ResponseMessage> deleteCustomer(Long id);
 
 }
